+++ src/Unchase.OData.ConnectedService/Provider.cs
@@ -29,7 +29,7 @@ namespace Unchase.OData.ConnectedService
                 BitmapSizeOptions.FromWidthAndHeight(32, 32)
             );
             CreatedBy = Constants.Author;
-            Version = new Version(0, 4, 0, 0);
+            Version = new Version(1, 1, 0, 0);
             Version = typeof(Provider).Assembly.GetName().Version;
             MoreInfoUri = new Uri(Constants.Website);
         }